Create main container and add style
THe body was given a red color so mimick the appearance of an etch a
sketch. Then, the main container for the 'pixels' was centered and
given a white color.

Both containers are flex, but the main container has the flex wrap
option active to allow the future containers to breake in a new line.