Create event listeners to draw in the screen
Two event listeners were added to each container performing as a pixel.
The first one calls a function named paintBlack() which adds a class
that possess a background black. This event takes place when the mouse
is hovering a pixel.

On the other hand, the second event listener calls the function
removeColor(), which removes the class added by the previous event
listener.