Create reset button
The button was declared in the HTML file.

In the Java file an event listener 'click' was added to the button,
which will call the function cleanSlate() that removes all of the
containers in the main one.

On the other hand, the variable pixel and the addition of event
listeners to each pixel were moved inside de createContainers functinon
so that every time it was reset, the pixels would change their color
again.