Create functions to paint RGB and black
RGBorBLACK function decides if the pixel shoudl be given a class among
green, red or blue; or if its color should have a black hue.

In order to get a random class among red, green or blue, Math random
was used to get integers from 0 to 2, which would be assigned a class.